separeate menu component and contents compomnent
