--- squre/src/Component/ClothsList/index.ts
@@ -1,2 +1,0 @@
-import ClothsList from "./ClothsList";
-export default ClothsList;
