+++ squre/src/Component/ClothsListAll/index.ts
@@ -1,0 +1,2 @@
+import ClothsListAll from "./ClothsListAll";
+export default ClothsListAll;
