+++ squre/src/Component/ClothsListBottom/index.ts
@@ -1,0 +1,2 @@
+import ClothsListBottom from "./ClothsListBottom";
+export default ClothsListBottom;
