+++ squre/src/Component/ClothsListOuter/index.ts
@@ -1,0 +1,2 @@
+import ClothsListOuter from "./ClothsListOuter";
+export default ClothsListOuter;
