+++ squre/src/Component/ClothsListTop/index.ts
@@ -1,0 +1,2 @@
+import ClothsListTop from "./ClothsListTop";
+export default ClothsListTop;
