--- squre/src/Component/Counter/Counter.tsx
@@ -1,14 +1,0 @@
-import React, { useState } from "react";
-
-const Counter = () => {
-  const [count, setCount] = useState(0);
-  return (
-    <div className="counter">
-      <h1> You clicked {count}</h1>
-      <button onClick={() => setCount(count + 1)}>up</button>
-      <button onClick={() => setCount(count - 1)}>down</button>
-    </div>
-  );
-};
-
-export default Counter;
