--- squre/src/Component/Counter/index.ts
@@ -1,2 +1,0 @@
-import Counter from "./Counter";
-export default Counter;
